changes: truncate comments when at str(Change)
Some people tend to write long commit comments for their work.
In fact, this goes back to the history of git, the purposes that led
to the original design by Linus.
Most commits of that project originate from emails sent to the LKML,
and the rule is that these emails should contain enough information
to justify the patch.
That's what git expects. A lengthy message with reasoning, tests, quotes
and discussion on the lines (many times one-liners).

But would you like to see all this spam in a log message? :P