git multi-poller: serialize calls to _doAddChange
The previous design would put all of them in a DeferredList, which meant
(by this twisted design) that all changes would start near-simultaneously
and then wait their results to be collected.
This, not only would try to send them all in parallel to the db (as was
noted a few days ago), but (in the openerp-buildbot subclass) could
trigger multiple secondary operations. There, we use an extra 'git log'
call, per change, in order to retrieve a detailed file list. Doing that
in parallel caused the algorithm to run out of OS open files.